requirements.txt
click==6.7
coverage==4.5.1
Flask==1.0.2
Flask-JWT==0.3.2
Flask-PyMongo==2.1.0
Flask-Script==2.0.6
Flask-Testing==0.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
PyJWT==1.4.2
pymongo==3.7.1
Werkzeug==0.14.1
